requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
albumentations==1.3.0
anyio==3.5.0
appnope==0.1.2
APScheduler==3.6.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
async-timeout==4.0.2
asyncpg==0.28.0
attrs==21.4.0
autopep8==1.6.0
awscli==1.25.80
Babel==2.9.1
backcall==0.2.0
bcrypt==4.0.1
beautifulsoup4==4.11.2
black==23.3.0
bleach==4.1.0
botocore==1.27.79
bs4==0.0.1
cachetools==4.2.2
Cerberus==1.3.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
click==8.0.3
codecov==2.1.13
colorama==0.4.4
coloredlogs==15.0.1
coverage==7.2.7
cryptography==39.0.1
cycler==0.11.0
Cython==0.29.35
databases==0.7.0
dataclasses-json==0.5.7
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.13
distro==1.8.0
docker==6.0.1
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
docutils==0.16
easydict==1.10
entrypoints==0.3
et-xmlfile==1.1.0
exceptiongroup==1.1.1
executing==0.8.2
fastapi==0.100.1
ffmpeg-python==0.2.0
filelock==3.12.0
Flask==2.0.2
flatbuffers==23.5.26
fonttools==4.34.4
free-proxy==1.1.0
frozenlist==1.3.3
future==0.18.3
greenlet==2.0.1
gunicorn==20.0.4
h11==0.14.0
httptools==0.6.0
humanfriendly==10.0
idna==3.3
imageio==2.30.0
iniconfig==2.0.0
insightface==0.7.3
ipykernel==6.7.0
ipython==8.0.1
ipython-genutils==0.2.0
isort==5.12.0
itsdangerous==2.0.1
jedi==0.18.1
Jinja2==3.0.2
jmespath==1.0.1
joblib==1.2.0
json5==0.9.6
jsonschema==3.2.0
jupyter-client==7.1.2
jupyter-core==4.9.1
jupyter-server==1.13.4
jupyterlab==3.2.8
jupyterlab-pygments==0.1.2
jupyterlab-server==2.10.3
kiwisolver==1.4.4
langchain==0.0.146
lazy_loader==0.2
Levenshtein==0.20.8
lxml==4.9.2
MarkupSafe==2.0.1
marshmallow==3.19.0
marshmallow-enum==1.5.1
matplotlib==3.5.2
matplotlib-inline==0.1.3
mistune==0.8.4
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==0.4.3
nbclassic==0.3.5
nbclient==0.5.10
nbconvert==6.4.1
nbformat==5.1.3
nest-asyncio==1.5.4
networkx==3.1
notebook==6.4.8
numexpr==2.8.4
numpy==1.24.3
onnx==1.14.0
onnxconverter-common==1.13.0
onnxmltools==1.11.2
onnxruntime-silicon==1.13.1
openai==0.19.0
openapi-schema-pydantic==1.2.4
opencv-python==4.7.0.72
opencv-python-headless==4.7.0.72
openpyxl==3.0.9
packaging==23.1
pandas==1.3.4
pandas-stubs==1.2.0.35
pandocfilters==1.5.0
parameterized==0.9.0
paramiko==3.0.0
parso==0.8.3
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
pipreqs==0.4.11
platformdirs==2.4.1
playsound==1.3.0
playwright==1.30.0
pluggy==1.0.0
pony==0.7.16
prettytable==3.7.0
prometheus-client==0.13.1
prompt-toolkit==3.0.26
protobuf==4.23.2
psutil==5.9.5
psycopg2-binary==2.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
pyasn1==0.4.8
pycodestyle==2.8.0
pycparser==2.21
pydantic==1.10.7
pydocstyle==6.3.0
pyee==9.0.4
Pygments==2.11.2
PyJWT==2.8.0
PyNaCl==1.5.0
pyparsing==3.0.7
pyrsistent==0.18.1
pyTelegramBotAPI==4.6.1
pytesseract==0.3.10
pytest==7.3.1
pytest-base-url==2.0.0
pytest-cov==4.1.0
pytest-playwright==0.3.0
python-dateutil==2.8.2
python-dotenv==0.19.2
python-slugify==8.0.0
python-telegram-bot==13.10
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.4.1
PyYAML==5.4.1
pyzmq==22.3.0
qudida==0.0.4
rapidfuzz==2.13.7
redis==4.1.2
replicate==0.8.1
requests==2.26.0
rsa==4.7.2
s3transfer==0.6.0
scikit-image==0.20.0
scikit-learn==1.2.2
scipy==1.10.1
Send2Trash==1.8.0
six==1.16.0
skl2onnx==1.14.1
sniffio==1.2.0
snowballstemmer==2.2.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.47
stack-data==0.1.4
starlette==0.27.0
store==2020.2.10
sympy==1.12
tenacity==8.2.2
terminado==0.13.1
testpath==0.5.0
text-unidecode==1.3
texttable==1.6.7
threadpoolctl==3.1.0
tifffile==2023.4.12
tk==0.1.0
toml==0.10.2
tomli==1.2.3
torch==2.0.1
tornado==6.1
tqdm==4.62.3
traitlets==5.1.1
typing-inspect==0.8.0
typing_extensions==4.5.0
tzdata==2021.5
tzlocal==4.1
unittest-tap-reporting==0.2
unittest-xml-reporting==2.1.0
urllib3==1.26.7
uvicorn==0.23.1
uvloop==0.17.0
waitress==2.1.2
watchfiles==0.19.0
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==0.59.0
websockets==11.0.3
Werkzeug==2.0.2
wget==3.2
wrapt==1.13.3
yarg==0.1.9
yarl==1.8.2
